[refac] do not use snprintf for debug logging and error messaging -
refactored the message building for error reporting, and all the
parser error logging to use the new dump facilities in c4core.

This avoids any risk of stack overflow from long file-names or long
source lines, which could happen because it was using a local stack
buffer. The new dump facilities bypass this problem because they:

  - are safer for building the error message since the overflow checks
    are part of the dump API
  - are safer for debug logging because:
      - arguments are dumped separately, without needing to serialize
        the full phrase into staging buffer
      - only non-string arguments use the staging buffer, and one
        at a time