[fix] Add missing Python runtime dep. on deprecation
The dependency on the PyPI “deprecation” package already appears in
requirements.txt, but it is not only a build-time dependency. The import
in api/ryml.i ends up in the generated ryml/ryml.py, so we need to add a
runtime dependency. This commit adds the necessary dependency via a new
“install_requires” argument in setup.py.