fix: add /Zc:__cplusplus for correct __cplusplus value in MSVC
When compiling with C++17 in MSVC (VS2022), error C2039: 'chars_format':
is not a member of 'std' occurs. This is due to the __cplusplus macro
always being set to 199711L, regardless of the specified C++ standard.
Consequently, conditionally included headers such as <charconv> are not
included.

Adding the /Zc:__cplusplus option ensures the __cplusplus macro reflects
the correct C++ standard. Please consider this change.